pyproject.toml
[project]
name = "nexen"
version = "0.1.0"
description = "NEXEN - Next-generation EXpert ENgine: Multi-Agent AI Research Assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "NEXEN Team"}
]
keywords = ["ai", "research", "multi-agent", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Agent Framework
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    
    # Multi-Model Gateway
    "litellm>=1.50.0",
    
    # Vector Store
    "qdrant-client>=1.12.0",
    
    # Configuration & Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    
    # Async & HTTP
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    
    # CLI
    "typer>=0.12.0",
    "rich>=13.9.0",
    
    # Utilities
    "tiktoken>=0.8.0",  # Token counting
    "tenacity>=9.0.0",  # Retry logic
    "structlog>=24.4.0",  # Structured logging
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
nexen = "nexen.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nexen"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
